Fix some issues with UUID case sensitivity
For the most part, we have standardized on storing and returning UUIDs
as strings with lower case hex digits in the format
xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx.

On the other hand, arguments to functions and methods should be case
insensitive to make it easy for end users.

This fixes some cases that didn't adhere to these principals.